chore(log_deployer): refactor packages in log-deployer #85
